+++ weblate/addons/git.py
@@ -215,7 +215,7 @@ class GitSquashAddon(BaseAddon):
                 # unless something is messed up)
                 try:
                     repository.execute(
-                        ["cherry-pick", commit, *gpg_sign],
+                        ["cherry-pick", "--empty=drop", commit, *gpg_sign],
                         environment={"WEBLATE_MERGE_SKIP": "1"},
                     )
                 except RepositoryError:
@@ -229,7 +229,7 @@ class GitSquashAddon(BaseAddon):
                         continue
                     try:
                         repository.execute(
-                            ["cherry-pick", other[0], *gpg_sign],
+                            ["cherry-pick", "--empty=drop", other[0], *gpg_sign],
                             environment={"WEBLATE_MERGE_SKIP": "1"},
                         )
                         handled.append(i)